requirements.txt
# ─────────────────────────────────────────────
#  Compliance Analysis API - Production Dependencies
#  Python 3.12+
# ─────────────────────────────────────────────

# ---------- API Framework ----------
fastapi==0.115.6
uvicorn[standard]==0.32.1

# ---------- Data Validation & Settings ----------
pydantic==2.10.4
pydantic-settings==2.7.0

# ---------- Configuration ----------
PyYAML==6.0.2
python-dotenv==1.0.1

# ---------- Document Ingestion ----------
PyMuPDF==1.24.4
python-docx==1.1.2

# ---------- Embeddings & Vector Store ----------
sentence-transformers==3.0.1
faiss-cpu==1.8.0

# ---------- LLM Adapters ----------
openai==1.59.6            # OpenAI + OpenAI-compatible (Ollama, vLLM, Together AI)
requests==2.32.3          # For Ollama HTTP API

# ---------- Utilities ----------
python-multipart==0.0.20  # multipart/form-data file uploads
pytest==8.2.2             # testing framework
